PlantPalMobile/backend_server/requirements.txt
Django>=4.2
djangorestframework
python-dotenv
supabase
requests
django-cors-headers
